refactor: update path references after compatibility/* rename
Updates every in-repo reference to the renamed compatibility harness
paths. No behaviour change — purely a path-rewriting commit that
follows the dir-rename commit so `go build ./...` and the CI lanes
work end-to-end after the move.

Touched categories:

  - Go import paths (compatibility/prometheus/shadow/cmd/shadow/*.go,
    compatibility/loki/cmd/loki-compliance-tester/main.go).
  - go.mod ignore directives.
  - Workflow scripts (compatibility/{prometheus,loki,tempo}/scripts/*.sh)
    and docker-compose contexts.
  - Dockerfile + Dockerfile.dockerignore for the tempo-compat driver.
  - Root .dockerignore + .markdownlintignore + .go-arch-lint.yml exempt
    paths.
  - Justfile recipes (compatibility / loki-compatibility /
    tempo-compatibility / compatibility-down / shadow-mode).
  - CI workflows ci.yml + chdb.yml + shadow-mode.yml path filters and
    exclusion globs.
  - CLAUDE.md architecture map + CONTRIBUTING.md allowlist note +
    docs/*.md (compatibility.md, tempo-compliance-plan.md,
    loki-compliance-plan.md, test-strategy.md, roadmap.md, coverage-
    baseline.md, pre-ga-*.md, compat-residual-audit-postE.md).
  - Comment-only mentions in seeder + driver source.

Leaves the CHANGELOG.md untouched (historical PR descriptions
referring to the old path) and the docs/roadmap.md external URL
pointing at BadLiveware/promshim-clickhouse (a different repo's
harness/ tree).
