+++ cmd/cerberus/main_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 // TestIsVersionFlag pins the argv shapes recognized by the
 // `--version` pre-flight. The cerberus container in
-// harness/prometheus-compliance/docker-compose.yml uses this exact path as
+// compatibility/prometheus/docker-compose.yml uses this exact path as
 // its docker healthcheck because the distroless runtime image has no
 // shell / wget / curl; a regression here would silently re-break the
 // compatibility lane's "container ... is unhealthy" failure mode
