+++ compatibility/loki/dataset_metadata.json
@@ -1,5 +1,5 @@
 {
-  "_comment": "Generated by reading harness/loki-compatibility/cmd/seed/main.go's deterministic fixture. The seeder pushes multi-label Loki streams for 13 services with JSON, logfmt, and unstructured log lines, anchored at 2026-05-11T00:00:00Z spanning 24 hours (1440 lines/service x 1min spacing). Each stream has structured metadata (detected_level) and log content includes error/debug/failed/refused keywords plus unwrappable numeric fields (duration, duration_ms, spans, status, bytes, size, streams).",
+  "_comment": "Generated by reading compatibility/loki/cmd/seed/main.go's deterministic fixture. The seeder pushes multi-label Loki streams for 13 services with JSON, logfmt, and unstructured log lines, anchored at 2026-05-11T00:00:00Z spanning 24 hours (1440 lines/service x 1min spacing). Each stream has structured metadata (detected_level) and log content includes error/debug/failed/refused keywords plus unwrappable numeric fields (duration, duration_ms, spans, status, bytes, size, streams).",
   "all_selectors": [
     "{cluster=\"cluster-0\", container=\"container-0\", datacenter=\"dc1\", env=\"production\", namespace=\"namespace-0\", pod=\"pod-0\", region=\"us-east-1\", service=\"web-server\", service_name=\"web-server\"}",
     "{cluster=\"cluster-0\", container=\"container-0\", datacenter=\"dc1\", env=\"production\", namespace=\"namespace-0\", pod=\"pod-1\", region=\"us-east-1\", service=\"database\", service_name=\"database\"}",
