+++ compatibility/prometheus/cerberus-test-queries.yml
@@ -1,7 +1,7 @@
 # Cerberus-side curated PromQL test queries for the compatibility harness.
 #
 # This file is a near-verbatim copy of
-# harness/prometheus-compliance/upstream/promql/promql-test-queries.yml
+# compatibility/prometheus/upstream/promql/promql-test-queries.yml
 # (prometheus/compliance @ 7e28242) with one edit:
 #
 # 1. The PromLabs "demo service" preamble (lines 1-31 of the upstream file)
