+++ compatibility/tempo/driver/Dockerfile.dockerignore
@@ -2,15 +2,15 @@
 #
 # BuildKit picks `<dockerfile>.dockerignore` over the repo-root
 # `.dockerignore` when both exist for the same Dockerfile. The root
-# `.dockerignore` excludes `harness/` (because Dockerfile.local builds
-# the cerberus binary and doesn't need it), but the tempo-compat
-# driver lives under `harness/tempo-compatibility/driver/` and DOES
-# need its own source visible in the build context. This file
-# narrows the exclusions accordingly.
+# `.dockerignore` excludes `compatibility/` (because Dockerfile.local
+# builds the cerberus binary and doesn't need it), but the tempo-compat
+# driver lives under `compatibility/tempo/driver/` and DOES need its
+# own source visible in the build context. This file narrows the
+# exclusions accordingly.
 #
 # Pattern: mirror the root .dockerignore exclusions for everything
-# unrelated to this driver, but DO NOT re-exclude `harness/` (so the
-# `COPY . .` in the Dockerfile picks up the driver sources).
+# unrelated to this driver, but DO NOT re-exclude `compatibility/` (so
+# the `COPY . .` in the Dockerfile picks up the driver sources).
 
 .git
 .github
@@ -51,11 +51,11 @@ docker-compose.yml
 test/e2e/k3s/
 test/e2e/grafana/
 
-# Narrow the harness/ exclusion: keep tempo-compatibility/driver/ in
-# the build context (needed for `go build`), but skip the heavy
+# Narrow the compatibility/ exclusion: keep compatibility/tempo/driver/
+# in the build context (needed for `go build`), but skip the heavy
 # upstream/ vendor + sibling harnesses to keep the image small.
-harness/prometheus-compliance/
-harness/loki-compatibility/
-harness/tempo-compatibility/upstream/
-harness/tempo-compatibility/reports/
-harness/tempo-compatibility/scripts/
+compatibility/prometheus/
+compatibility/loki/
+compatibility/tempo/upstream/
+compatibility/tempo/reports/
+compatibility/tempo/scripts/
